requirements.txt
amqp==2.2.2
apipkg==1.4
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
bs4==0.0.1
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==1.9
Django==1.11
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-cors-headers==2.2.0
django-redis==4.9.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
execnet==1.5.0
Fabric3==1.14.post1
html5lib==1.0.1
idna==2.6
JPype1==0.6.3
kombu==4.1.0
konlpy==0.5.1
lxml==4.1.1
more-itertools==4.2.0
nltk==3.3
numpy==1.14.0
pandas==0.22.0
paramiko==2.4.1
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.3
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.4
PyNaCl==1.2.1
pytest==3.6.1
pytest-cov==2.5.1
pytest-django==3.3.0
pytest-forked==0.2
pytest-xdist==1.22.2
python-dateutil==2.6.1
pytz==2017.3
raven==6.9.0
requests==2.18.4
scipy==1.0.0
six==1.11.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
xlrd==1.1.0